Add: Player prefab variant
I disabled the player prefab we had in the scene, created a prefab variant from it and placed it in the scene. I also changed the layer of the PPvolume to the newly created one from last push and added a volume component while disabling the post processing volume component. In the main camera of the player prefab variant, I enabled post processing and in environment; changed the layer mask to mask the one that the PPvolume is using.